PrayerFunction/src/main/java/org/jhaughton/orison/PrayerApp.java
package org.jhaughton.orison;

import com.amazonaws.services.lambda.runtime.Context;
import com.amazonaws.services.lambda.runtime.RequestHandler;
import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent;
import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent;
import com.google.gson.Gson;
import org.jhaughton.orison.model.Organisation;
import org.jhaughton.orison.model.Prayer;

import java.util.HashMap;
import java.util.Map;

/**
 * Handler for requests to Lambda function.
 */
public class PrayerApp implements RequestHandler<APIGatewayProxyRequestEvent, APIGatewayProxyResponseEvent> {

    public APIGatewayProxyResponseEvent handleRequest(final APIGatewayProxyRequestEvent input, final Context context) {
        Map<String, String> headers = new HashMap<>();
        headers.put("Content-Type", "application/json");
        headers.put("X-Custom-Header", "application/json");

        APIGatewayProxyResponseEvent response = new APIGatewayProxyResponseEvent()
                .withHeaders(headers);
        Organisation organisation = new Organisation("Test Org");
        Prayer prayer = new Prayer("Test Content", organisation);
        Gson gson = new Gson();
        String output = gson.toJson(prayer);

        return response
                .withStatusCode(200)
                .withBody(output);
    }
}
